chore: move types to component folders
