+++ packages/react/src/components/Storage/FileUploader/FileUploader/utils.ts
@@ -1,0 +1,4 @@
+import type { UploadTask } from '@aws-amplify/storage';
+
+export const isUploadTask = (value: unknown): value is UploadTask =>
+  typeof (value as UploadTask)?.resume === 'function';
