+++ packages/react/src/components/Storage/FileUploader/UploadDropZone/types.ts
@@ -1,0 +1,6 @@
+import { DragActionHandlers } from '../hooks/useFileUploader/types';
+
+export interface UploadDropZoneProps extends DragActionHandlers {
+  children?: React.ReactNode;
+  inDropZone?: boolean;
+}
