+++ packages/react/src/components/Storage/FileUploader/UploadMessage/types.ts
@@ -1,0 +1,7 @@
+import { FileState } from '../types';
+
+export interface UploadMessageProps {
+  fileState: FileState;
+  errorMessage: string;
+  percentage?: number;
+}
