+++ packages/react/src/components/Storage/FileUploader/UploadPreviewer/types.ts
@@ -1,0 +1,14 @@
+import { FileStatuses } from '../types';
+
+export interface UploadPreviewerProps {
+  aggregatePercentage: number;
+  children?: React.ReactNode;
+  dropZone: React.ReactNode;
+  fileStatuses: FileStatuses;
+  isLoading: boolean;
+  isSuccessful: boolean;
+  hasMaxFilesError: boolean;
+  maxFileCount: number;
+  onClear: () => void;
+  onFileClick: () => void;
+}
