+++ packages/react/src/components/Storage/FileUploader/UploadTracker/types.ts
@@ -1,0 +1,18 @@
+import { FileState } from '../types';
+
+export interface UploadTrackerProps {
+  errorMessage: string;
+  file: File;
+  fileState: FileState;
+  hasImage: boolean;
+  name: string;
+  onCancel: () => void;
+  onCancelEdit?: () => void;
+  onPause: () => void;
+  onResume: () => void;
+  onSaveEdit: (value: string) => void;
+  onStartEdit: (event: React.MouseEvent<HTMLButtonElement, MouseEvent>) => void;
+  percentage: number;
+  isResumable?: boolean;
+  showImage: boolean;
+}
